first refactorings, not runnable yet
