+++ apiserver/pom.xml
@@ -16,7 +16,7 @@
         <dependency>
             <groupId>org.jboss.resteasy</groupId>
             <artifactId>resteasy-cache-core</artifactId>
-            <version>3.7.0.Final</version>
+            <version>4.1.0.Final</version>
             <exclusions>
                 <exclusion>
                     <groupId>org.infinispan</groupId>
