+++ backupservice-defintion/pom.xml
@@ -22,7 +22,7 @@
         <dependency>
             <groupId>org.jboss.resteasy</groupId>
             <artifactId>resteasy-multipart-provider</artifactId>
-            <version>3.7.0.Final</version>
+            <version>4.1.0.Final</version>
         </dependency>
     </dependencies>
