+++ configurationrepository-definition/pom.xml
@@ -17,7 +17,7 @@
         <dependency>
             <groupId>org.jboss.resteasy</groupId>
             <artifactId>resteasy-jackson2-provider</artifactId>
-            <version>3.7.0.Final</version>
+            <version>4.1.0.Final</version>
             <scope>provided</scope>
             <exclusions>
                 <exclusion>
